requirements.txt
discord.py
requests
retry